requirements-scraper.txt
playwright>=1.40.0
python-dotenv>=1.0.0
streamlit>=1.31.0
pandas>=2.1.0
plotly>=5.18.0
openpyxl>=3.1.0
